Move HTML page handlers into HtmlServer class
Introduce HtmlServer class in html.cpp/html.h to encapsulate
all HTML web UI page handlers previously scattered as free
functions in cscout.cpp.
